Updating code to new discord.js
Update to discord.js 12.5.1 and add sharding management

+++ commands/fun/say.js
@@ -8,8 +8,9 @@ module.exports = {
     example: ['Hello, World!']
 };
 
-const { RichEmbed } = require('discord.js');
 module.exports.run = async (bot, message, args, suffix) => {
+    if(!args[0]) return;
+
     message.delete();
 	message.channel.send(suffix);
 }
